requirements.txt
amqp==2.6.1
asgiref==3.2.10
attrs==20.1.0
Automat==20.2.0
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
constantly==15.1.0
cryptography==3.0
cssselect==1.1.0
Django==3.1
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-filter==2.4.0
djangorestframework==3.11.1
httpie==2.2.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
itemadapter==0.1.0
itemloaders==1.0.2
jmespath==0.10.0
kombu==4.6.11
lxml==4.5.2
parsel==1.6.0
Pillow==7.2.0
Protego==0.1.16
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
Pygments==2.7.0
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pytz==2020.1
queuelib==1.5.0
redis==3.5.3
requests==2.24.0
Scrapy==2.3.0
scrapy-djangoitem==1.1.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
Twisted==20.3.0
urllib3==1.25.10
vine==1.3.0
w3lib==1.22.0
zope.interface==5.1.0
